+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/FileSystemModule.kt
@@ -29,6 +29,27 @@ class FileSystemModule : Module() {
 
   private val downloadStore = DownloadTaskStore()
 
+  private fun writeToFile(
+    file: FileSystemFile,
+    content: Either<String, TypedArray>,
+    options: WriteOptions?
+  ) {
+    val append = options?.append ?: false
+    if (content.`is`(String::class)) {
+      content.get(String::class).let {
+        if (options?.encoding == EncodingType.BASE64) {
+          file.write(Base64.decode(it, Base64.DEFAULT), append)
+        } else {
+          file.write(it, append)
+        }
+      }
+    } else if (content.`is`(TypedArray::class)) {
+      content.get(TypedArray::class).let {
+        file.write(it, append)
+      }
+    }
+  }
+
   @RequiresApi(Build.VERSION_CODES.O)
   override fun definition() = ModuleDefinition {
     Name("FileSystem")
@@ -138,22 +159,12 @@ class FileSystemModule : Module() {
         file.create(options ?: CreateOptions())
       }
 
-      Function("write") { file: FileSystemFile, content: Either<String, TypedArray>, options: WriteOptions? ->
-        val append = options?.append ?: false
-        if (content.`is`(String::class)) {
-          content.get(String::class).let {
-            if (options?.encoding == EncodingType.BASE64) {
-              file.write(Base64.decode(it, Base64.DEFAULT), append)
-            } else {
-              file.write(it, append)
-            }
-          }
-        }
-        if (content.`is`(TypedArray::class)) {
-          content.get(TypedArray::class).let {
-            file.write(it, append)
-          }
-        }
+      AsyncFunction("write") Coroutine { file: FileSystemFile, content: Either<String, TypedArray>, options: WriteOptions? ->
+        writeToFile(file, content, options)
+      }
+
+      Function("writeSync") { file: FileSystemFile, content: Either<String, TypedArray>, options: WriteOptions? ->
+        writeToFile(file, content, options)
       }
 
       AsyncFunction("text") { file: FileSystemFile ->
@@ -261,10 +272,16 @@ class FileSystemModule : Module() {
       Constructor { file: FileSystemFile, mode: FileMode? ->
         file.openHandle(mode)
       }
-      Function("readBytes") { fileHandle: FileSystemFileHandle, bytes: Long ->
+      AsyncFunction("readBytes") { fileHandle: FileSystemFileHandle, bytes: Long ->
         fileHandle.read(bytes)
       }
-      Function("writeBytes") { fileHandle: FileSystemFileHandle, data: ByteArray ->
+      Function("readBytesSync") { fileHandle: FileSystemFileHandle, bytes: Long ->
+        fileHandle.read(bytes)
+      }
+      AsyncFunction("writeBytes") { fileHandle: FileSystemFileHandle, data: ByteArray ->
+        fileHandle.write(data)
+      }
+      Function("writeBytesSync") { fileHandle: FileSystemFileHandle, data: ByteArray ->
         fileHandle.write(data)
       }
       Function("close") { fileHandle: FileSystemFileHandle ->